requirements.txt
Django
djangorestframework
drf-yasg
mysqlclient